+++ gcs/CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Next
 1. Enable move by default
+2. [Feature] Add support for cloud logging and prefix invocation ID to logs for correlation.
 
 ## 3.0.11 - 2025-09-09
 1. PR #1484 - [Feature] Enabled write checksum by default